requirements.txt
psycopg2==2.8.4
pyotp==2.3.0
django-allauth==0.32.0
django-rest-auth==0.9.5
djangorestframework-jwt==1.11.0
djoser==2.0.3
aioredis==1.3.0
amqp==2.5.2
asgiref==3.2.3
asn1crypto==1.2.0
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
backports.csv==1.0.7
base58check==1.0.2
billiard==3.6.1.0
celery==4.3.0
certifi==2019.11.28
cffi==1.13.2
channels==2.3.1
channels-redis==2.4.1
chardet==3.0.4
coinaddr==1.0.1
constantly==15.1.0
cryptography==2.8
daphne==2.4.0
defusedxml==0.6.0
django==2.2.7
django-autoslug==1.9.6
django-channels==0.7.0
django-countries==5.5
django-crequest==2018.5.11
django-crispy-forms==1.8.1
django-environ==0.4.5
django-extensions==2.2.5
django-filter==2.2.0
django-jet==1.0.8
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
dnspython==1.16.0
email-validator==1.0.5
et-xmlfile==1.0.1
gunicorn==20.0.4
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
iyzipay==1.0.32
jdcal==1.4.1
kombu==4.6.6
lazy-object-proxy==1.4.3
markdown==3.1.1
mccabe==0.6.1
msgpack==0.6.2
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==3.0.2
pep8==1.7.1
Pillow==6.2.1
# psycopg2==2.8.4
# psycopg2-binary==2.8.4
pycountry==19.8.18
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.11
pylint-plugin-utils==0.6
pysha3==1.0.2
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
qrcode==6.1
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
Twisted==19.10.0
txaio==18.8.1
typed-ast==1.4.0
urllib3==1.25.7
vine==1.3.0
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.7.1
django-widget-tweaks==1.4.5
django_countries==5.5
django-ckeditor==5.9.0


